fix bug: template compile failed (got "HtmlWebpackPlugin Error: No source available") in webpack watch mode.
example webpack.config.js:

```js
module.exports = {
    entry: {a: "./src/a.js", b: "./src/b.js"}
    debug: true,
    output: {
        path: "./var",
    },
    plugins: [
        new HtmlWebpackPlugin({filename: "a.html", template: "./src/template.html"}),   // a.html
        new HtmlWebpackPlugin({filename: "b.html", template: "./src/template.html"}),   // b.html
    ]
};
```

These html file use the same template.
In webpack watch mode, at first, they all can successful compile, but now if you edit `./src/template.html`, only `a.html` can successful compile again, `b.html`'s content will change to "HtmlWebpackPlugin Error: No source available".

I think this is because webpack cached the module (a `NormalModel`) of `a.html`, and when `b.html` going to compiling, it reused the cached one.
In the first compilation, everything is ok, but in second compilation, `b.html` was compiled **before** `a.html` (don't know why),
it used `a.html`'s module before the module fully initialized. So the `_source` property of this module is `null`, and the compilation of `b.html` was failed.

The trace is:

```
Compiler.runAsChild -> Compiler.compile -> this.applyPluginsParallel("make") -> Compilation.addEntry -> Compilation._addModuleChain -> this.addModule(module)
```

To resolve this problem, I created a queue: `compileQueue`.
If two or many child compiler use the same template, the second one will wait for first one compile complete, then start it's own compilation.

I don't know if this is a good solution, or if this is the real reason of causing this bug, what do you think?