add documentation & an example for using one entry point per page
Improves README.md to mention the `chunks` option in sections that might
be of interest to users who have multiple entry-points and/or HTML
files.

Also adds a new example demonstrating how multiple entry points AND
multiple pages may be used together, alongside the CommonsChunkPlugin.

Based on an example written by @simshanith in #218.