Change in-line spelling (#777)
I've searched for inlining solution by Ctrl+F'ing `inline`. Went for #104 and found last comment by luck. Then only found that this plugin was already mentioned in README. Spelling `in-line` is a bit strange, in my opinion. I understand that this is a smallish change, so feel free to decline changes. :v: