chore: No longer leave a comment when issues/PRs are locked
Since it creates too much email notification noise, and the reason
for locking should be clear enough on its own (its performed by a
username of `lockbot` and the status of the issue/PR is set to
`"resolved"`).

I decided against adding a label, since it will only end up adding
noise to the labels list and issue/PR search results that adds little
value other than as a workaround for inability to set a custom GitHub
lock reason.

Fixes #1085.