Add stacktrace to error message from childCompilation.errors
I was had an error that was thrown in my webpack.config.ts that was causing childCompilation to fail. But I didn't know it was an error thrown in my webpack.config.ts because child-compiler.js didn't include the stack trace in the error message.

The stack trace I got ended at line 122 in node_modules\html-webpack-plugin\lib\child-compiler.js. So I fiddled with that to make it output the inner error stack trace and then I got the actual place of the error.