feat: Support type=module via scriptLoading option
Add third option, `'module'`, to the `scriptLoading` option.

This partially fixes issue #1663, which is also where this solution to
the problem was proposed. Although this patch doesn't give a way to
freely set `type` attribute to any value, setting `scriptLoading:
"module" now makes the scripts have attribute `type="module"`, which was
the original motivation behind the issue.